Move api module to package root
